requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
pydantic
pydantic-settings
python-jose
passlib[bcrypt]
email-validator
pytest
pytest-asyncio
httpx
websockets
google-genai
apscheduler